refactor: rename `ui/utils` to `ui/ui_utils` for clarity
